+++ pdf/lib/src/widgets/grid_paper.dart
@@ -270,45 +270,45 @@ class GridPaper extends SingleChildWidget {
       n++;
     }
 
-    if (border.left.style != BorderStyle.none) {
-      BoxBorder.setStyle(context, border.left.style);
+    if (border.left.style.paint) {
+      border.left.style.setStyle(context);
       context.canvas
         ..setStrokeColor(border.left.color)
         ..setLineWidth(border.left.width)
         ..drawLine(box!.left + margin.left, box!.top, box!.left + margin.left,
             box!.bottom)
         ..strokePath();
-      BoxBorder.unsetStyle(context, border.left.style);
+      border.left.style.unsetStyle(context);
     }
-    if (border.right.style != BorderStyle.none) {
-      BoxBorder.setStyle(context, border.right.style);
+    if (border.right.style.paint) {
+      border.right.style.setStyle(context);
       context.canvas
         ..setStrokeColor(border.right.color)
         ..setLineWidth(border.right.width)
         ..drawLine(box!.right - margin.right, box!.top,
             box!.right - margin.right, box!.bottom)
         ..strokePath();
-      BoxBorder.unsetStyle(context, border.right.style);
+      border.right.style.unsetStyle(context);
     }
-    if (border.top.style != BorderStyle.none) {
-      BoxBorder.setStyle(context, border.top.style);
+    if (border.top.style.paint) {
+      border.top.style.setStyle(context);
       context.canvas
         ..setStrokeColor(border.top.color)
         ..setLineWidth(border.top.width)
         ..drawLine(
             box!.left, box!.top - margin.top, box!.right, box!.top - margin.top)
         ..strokePath();
-      BoxBorder.unsetStyle(context, border.top.style);
+      border.top.style.unsetStyle(context);
     }
-    if (border.bottom.style != BorderStyle.none) {
-      BoxBorder.setStyle(context, border.bottom.style);
+    if (border.bottom.style.paint) {
+      border.bottom.style.setStyle(context);
       context.canvas
         ..setStrokeColor(border.bottom.color)
         ..setLineWidth(border.bottom.width)
         ..drawLine(box!.left, box!.bottom + margin.bottom, box!.right,
             box!.bottom + margin.bottom)
         ..strokePath();
-      BoxBorder.unsetStyle(context, border.bottom.style);
+      border.bottom.style.unsetStyle(context);
     }
 
     context.canvas.restoreContext();